% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce}
\alias{coalesce}
\alias{\%||\%}
\title{Coalesce two R objects}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Any R object, possibly \code{NULL}}

\item{y}{Any R object, possibly \code{NULL}}
}
\description{
I first saw this function in
\href{https://youtu.be/7oyiPBjLAWY}{a talk by Jenny Bryan at UseR! 2018}.
I hear it referred to as "coalesce", most likely named after the SQL
function. I later found it out it was actually included in \code{rlang}, but I
tend to keep my own implementation of the function (and its unit tests)
rather than introduce a dependency.
}
\examples{
3 \%||\% NULL
NULL \%||\% "red panda"
NULL \%||\% NULL \%||\% 42

}
