% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-manipulation.R
\name{make_url_friendly}
\alias{make_url_friendly}
\title{Convert an R object into a string that can be included in a URL}
\usage{
make_url_friendly(input)
}
\arguments{
\item{input}{Logical or character.}
}
\value{
Character.
}
\description{
Before a character can be used as part of a html, spaces must be converted to
"\%20", and R logicals must be converted to "true"/"false".
}
